+++ src/pages/ProductDetail/ProductImage/ProductImage.js
@@ -1,6 +1,6 @@
 import React, { useEffect, useState } from 'react';
 import { useLocation } from 'react-router-dom';
-import BASE_URL from '../../config';
+import BASE_URL from '../../../config';
 import Slider from 'react-slick';
 import styled from 'styled-components';
 import 'slick-carousel/slick/slick.css';
