+++ src/pages/ProductDetail/ProductOtherInfo/ProductOtherInfo.module.scss
